requirements.txt
shiny
plotly
pandas
requests
python-dotenv
shinyswatch
shinywidgets>=0.3.2